[project structure dialog] fix editing dependencies after pressing 'Apply' (IDEA-264113, EA-253839)
Items from the table model may be used for modifications, so we need to take them from a modifiable model. If the table is updated inside 'rootsChanged' synchronously, it will cause problems because the new modifiable model will be created in the middle of ModulesConfigurator.apply, postponing via 'invokeLater' fixes the problem.

GitOrigin-RevId: c38f2cf6440740acfb9ff7c0f4917427b6f53692