PY-47633 Update PropertiesManager to process imports in all definitions of a property
It turned out that saving/restoring imports in MethodsManager and
PyClassRefactoringUtil duplicated the logic already present in
TypeSafeMovingStrategy. Removing the redundant operations didn't immediately
work for refactorings over properties, however, for two reasons:
- PropertiesManager didn't return all components of a property definition
as elements which references need to be remembered (only its getter).
- TypeSafeMovingStrategy didn't handle references to other moved elements
and tried to re-create imports for already invalid PSI.

GitOrigin-RevId: f27d0dc192002b0fc42f923cd8b25664bd14257e