A ConsoleView is disposed if its editor is disposed
If the ConsoleView is disposed from a background thread, it may be in a state when the UI part of disposeEditor has already executed but the remaining part in the background thread has not resumed yet. In this state, if we try to do something with the console in EDT, we could get an "editor disposed" exception.

GitOrigin-RevId: ca92d9b68787576ba45dfc423b39f2adc95db7e9