Fix MultiModuleMoveTestGenerated.testMoveFromJvmModuleToJsModule_MoveFromJvmModuleToJsModule
KTIJ-5661

Test is about not being possible to invoke "Move" refactoring on Kotlin class from JVM module to
JS module if class mentions @JvmStatic annotation somewhere inside. @JvmStatic annotation is
present in kotlin-stdlib-common (that's why it's possible to resolve it in JS module) but it's
annotated with @OptionalExpectation annotation.

From @OptionalExpectation annotation documentation:
```
Marks an expected annotation class that it isn't required to have actual counterparts in all platforms.
```

GitOrigin-RevId: 89f80454cc9601b005849e8d08fd7d296782cd1d