[EA-260251] IAE: Internals.checkLength: fixes after review
This patch eliminates the StreamEx#zip method's invocations as per the review request and the type of an element from a throws list is calculated via PsiElementFactory instead of calling PsiReferenceList#getReferencedTypes. This patch also adds logging if an inconsistency between Stub and PSI is detected in order to better understand what causes the inconsistency.

GitOrigin-RevId: c1f9bfa5f88292496cb9e3ff8f7a7a1962d3a8b1