[workspace model] don't delete '__external-system-id' attribute from libraries, artifacts and facets under .idea
We need to store additional attributes in xml files to distinguish facets and libraries imported from external systems even if "store externally" option is switched off, in order to properly show warning when user tries to modify them (IDEA-264750) and to simplify code which replaces these entities during reimport. However if we add such attributes in a new version, this will modify configuration files which are shared via VCS, and if user commits these changes, and then somebody opens the project in a previous version of the IDE, it'll remove the attributes. In order to reduce annoyance we can ensure that IDE won't remove these attributes if they are added (but ignore them for now).

GitOrigin-RevId: 7eebab0c774b7374490dcfb315af9893c6bd7c67