[vcs-log] remove cleaning up cache files for old storage layout
In 9ffcf01c842db5f393c71d4dc63cacf72a837ae9 storage layout was changed to group cache files by project. A call for deleteAllFilesStartingWith was added to remove files in the old layout (not grouped by project). This is no longer needed, as it seems reasonable to assume that all of the caches were migrated since then.

GitOrigin-RevId: 10d65e36b9cc0269df2f9b19dab82a1f95fdee7a