[vcs-log] add consistency check for index storage
When the set of indexed commits is not empty, check that all of the indexes are not empty as well. If they are, throw an exception to force full reindex.

This commit almost is the same as reverted f28ffbbaa180b7c2bfd0a84ca452ca97ae5c4e2c. The difference is in PersistentSet#isEmpty implementation, which uses PersistentEnumeratorBase#traverseAllRecords instead of PersistentEnumeratorBase#iterateData (not supported when key storage is inlined).

IDEA-265042

GitOrigin-RevId: f9312b806a99fe950188c35f36bef58c29603a07