Remove unused parameter from CommonJavaTargetTestBase.doTestJUnitRunConfiguration() test method
`SearchForTestsTask.startSearch()` executes the logic either under the process indicator or in the same thread when it runs in the test mode and the specific key (`CONNECT_IN_UNIT_TEST_MODE_PROPERTY_KEY`) is absent. The key was utilized in `CommonJavaTargetTestBase` class (formerly `JavaTargetTestBase`), where some tests switched it on and some left it turned off. However all tests in `CommonJavaTargetTestBase` inheritors pass with the key on, so the parameter `runWithConnectInUnitTestMode` is removed.

GitOrigin-RevId: b41481c16c9c2a1afbb7c1292a048d27764b1f3b