[workspace model] fire 'rootsChanged' event if a parent directory for a library JAR file is created (IDEA-265594)
In that case VFS won't send events about files under directory, so we need to search for roots under the created directory in JAR filesystem. Such cases often happen after updating a Maven library to a new version, because the JAR file for a new version is located in a new subdirectory.

This change also fixes problem with directory roots which have '.jar' extension.

GitOrigin-RevId: ac781cd2f6029553c3c45a0ebbd5b35d126489d6