IDEA-224332 Replace line comment before a method with a Javadoc comment
Adds a new intention with the name "replace with javadoc" that converts end-of-line comments and c-style block comments with javadoc comments.

The idea is to grab all the comments (left and right siblings) that belong to a PsiMember, squash them together and make a javadoc out of their combined content.

GitOrigin-RevId: b93bbab4a31d318c8b6e0557a7c6f88622f4abac