fix update of 'persistent' highlighters on document changes
There exist two problems now with 'persistent' highlighters, both related to storing of line number as a field in PersistentRangeHighlighterImpl:
1) That field isn't updated on document change happening far from current highlighter location. Such a change leaves the line number out of sync
   with actual highlighter location (defined by its offsets), and another update, performed via diff mechanism (e.g. for a whole text replacement)
   might reposition the highlighter incorrectly.
2) For document changes in the vicinity of highlighter that field is updated only for one highlighter in a tree node (if there are several persistent
   highlighters with identical position and flags). If that highlighter is disposed, other highlighters might be updated incorrectly later,
   as the line number is out of sync for them.

The described problems are fixed by not storing line number with the highlighter - its value can be restored from the document event when needed.
This requires parsing of old document content, but it's done anyway to build the diff of changes used for highlighter update.

hopefully, this will fix EA-257554 (T: ErrorStripeMarkerImpl.validateState)

GitOrigin-RevId: 04e19da35a0c6ee948da4ca11f72b560d9449778