Kotlin evaluator: use reflection for calling private declarations (KTIJ-5622)
Similarly to the evaluator for Java, Kotlin evaluator used to rely on the MagicAccessorImpl JDK class for accessing private declarations from evaluated code blocks. However, the class was made internal in JDK 9, and accessing it involved so much magic and shamanism so it turned to be impractical to stick with it.

Then Java debugger switched to pre-processing of the evaluated code fragment PSI, specifically, on replacing all method calls and field accesses with reflection. It worked reasonably well, so Kotlin debugger decided to follow the same way.

Yet, it's much more tricky to modify Kotlin PSI. Instead, we decided to process the generated bytecode instructions. Now, the patcher supports ordinary method calls (invoke-virtual and invoke-static) and field accesses (get/put-field, get/put-static).

Known issues:
- Constructor calls are not supported (KTIJ-7649)
- Private classes can't be initiated (KTIJ-7650)
- Methods with inline classes in signatures (KTIJ-7652)

GitOrigin-RevId: 01b39079232d1cc2a09f5179c0ebb00b34976744