Assorted fixes to inspection gadget HTML snippets.
Added additional context to PointlessBitwiseExpression.html

- Updated the initial description.
- Added comments to the first and second examples to explain why these
    expressions are pointless, as especially for new developers, this
    may be confusing.
- Added an underscore to the hexadecimal value of Integer's max value
    to make it clearer to read, which may help those with visual
    impairments or who rely on a screenreader.

Moved comments in SerializableRecordContainsIgnoredMembers.html

The comments that previously resided at the end of the lines have been moved to
the top of the items that they are discussing. This reduces the amount of content
that either wraps or requires horizontal scrolling in the settings panel, which
improves readability a little.

Fixing less/greater symbols in InspectionGadgets

Fixes multiple cases where ">" and "<" may be interpreted as HTML
markup by mistake. This is currently present in IDEA Community 2020.3,
and appears to have not been fixed within these files since.

Fixing inconsistent file endings on edited files.

Code review formatting.

GitOrigin-RevId: 2fe199296cec57f584f4e6dc965e5cf521591f93