remove `PerformInBackgroundOption` usage from `BackgroundableProcessIndicator`
.. since all backgroundable progresses start in background right away (IDEA-234632).

The `background` method is called when the user clicks "Minimize" in the progress window.
The progress window is still shown when this indicator is used on a welcome screen, in such case
`PerformInBackgroundOption#shouldStartInBackground` was also ignored which resulted in showing a progress window.
Calling `PerformInBackgroundOption#processSentToBackground` after ignoring `PerformInBackgroundOption#shouldStartInBackground`
resulted in the following behaviour: the user backgrounds an unconditionally shown modal progress on a welcome screen,
and then the same progress might've never appeared as a window in a regular project frame if `processSentToBackground` saved the state.

^ This means `processSentToBackground()` call can be removed from `#background` method
for consistency with removal of `shouldStartInBackground()` call.

GitOrigin-RevId: 53cb99222f679f499de1dc6ad49303d65d03cc37