PY-46321 Disabled the code formatter for PyDocstring language
PyDocstring is based on Python, but its custom tokens `>>>` and `...` are registered just as whitespace. These tokens are ignored by the Python code formatter, which results in exceptions from the formatting model, which doesn't cover the whole text range.

We could either fix the formatting model or just disable formatting for PyDocstring. I prefer the latter for now, since the current approach with PyDocstring being a dialect of Python results in many subtle errors. It's better to have a workaround for now and maybe re-write doctest support later.

GitOrigin-RevId: 68b9a57650631f5d7245c85ada294af85c0aad44