PY-47905 Disable completion for non-imported modules and their attributes in console
Pydev console has an internal representation very different from a regular file
in the editor -- each line is effectively a separate file. In particular, standard
mechanisms of resolve over PSI tree, like PyResolveUtil.resolveLocally(), are not
applicable there, and, thus, every reference is considered unresolved by these terms.
As a result, the new completion for not imported modules and packages gets
triggered for many irrelevant cases. While it's not clear how to make it work
properly in Python console, the easiest workaround is to disable this feature
there altogether for the time being.

GitOrigin-RevId: 6b720684a116c40af459b089ddf85b6061586a8e