Fix restarter and libnst64 universal binaries for Mac
With commit d0480b907dd63e48f8826c810138e21acd16c8fb only fsnotifier
was updated to a universal binary. With the current Xcode toolchain,
setting C as the project language in CMakeLists.txt results in
"-arch x86_64 -arch arm64" being passed correctly to clang, whereas
with OBJC this is not the case.

Bug: 185748850
Test: manual
Change-Id: Ifeef4a09cd07fad08ba424855969bc9387848a21