IDEA-244182 JCEF heavyweight popup closes too early or always stays on top [Linux]
On Linux, when JCEF is shown in a heavyweight popup, the WINDOW_GAINED_FOCUS event dispatched to the frame has correct opposite, but WINDOW_DEACTIVATED has null opposite. This causes IdePopupManager to close the popup.

It could be not trivial to fix the opposite in that case. As an alternative, we correct the fix that brought WINDOW_DEACTIVATED event to here - IDEA-185353. It was fixing the problem on macOS and so we apply it to macOS only (where IDEA-244182 is not reproduced).

GitOrigin-RevId: ced0fdf3a48415ed11c358315682c44ae49d8551