Stop freeze checker from blocking UI thread
This is a workaround for IDEA-253433. PerformanceWatcher
writes freeze report to disk on UI thread which creates
additional UI freeze.

By removing call to Future.get(), UI thread is not blocked anymore.
This change can impact registering of any future freezes - if it takes
more than 5 seconds to write all files to disk and there is subsequent
freeze to register, it may be either ignored or uiFreezeStarted()
event will be reported later.

Fixes: 166548413
Test: N/A
Change-Id: I728f2dcb46eace5a7bb5a75c7355be9261c69716