[git-index] always show branch in tab title for single-root projects
Since 253f8dd6aec82e3781cbca11edb780574f16f5f8 repository grouping was enabled by default in order to allow committing vcs roots separately. This was done regardless of the number of roots. Because of this, additional check for roots number is required to show branch in tab title.

IDEA-267252

GitOrigin-RevId: ab6976a5bb63d30a77061b782e95ed3b649aa1f4