[comments][foundByPropertyTesting] False positive ReplaceWithJavadocIntention's action registration
ReplaceWithJavadocIntention registers its action when an end-of-line or block comment is above a class member. Java PSI tree pushes such comments inside a PsiMember node, that is why it was safe to check if a comment's parent is an instance of PsiMember. It didn't take into account that when a comment is inside an anonymous class is not actually above the anonymous class, but it still inside a PsiAnonymousClass node. This patch excludes nodes of PsiAnonymousClass from being examined in order to exclude the false positive action registration.

GitOrigin-RevId: 5e7bf3818ca3e3b4f94031e0f297f187dbdba2ce