[vcs-log] fix stuck cursor in the table
The problem with cursor happens because of the conflict between a cursor set by the current VcsLogCellController and a drag cursor displayed on the column border by the InvisibleResizableHeader. The order of events leading to the problem is as follows:
1. Custom cursor is set by VcsLogGraphTable. Old cursor is saved in the "myLastCursor" field.
2. Mouse moves to the column border.
3. InvisibleResizableHeader changes cursor to the drag cursor.
4. Mouse moves to the other column with default cursor.
5. VcsLogGraphTable tries to restore the cursor, but can not do so as it was changed. Field "myLastCursor" is not cleared.

Before this commit, this sequence caused the table to be stuck without the ability to set custom cursor again. This commit clears "myLastCursor" field even if it is unable to restore it.

This fix does not solve cursor problems completely, as the interplay between two pieces of code working with cursor is still rather complex and leads to other issues.

GitOrigin-RevId: 1edd3da8b19afcc12037b9a9e1071c4e09db16af