[workspace model] remove specified dependency only if there are several equal dependencies (IDEA-267246, IDEA-256694)
If there are several instances of OrderEntry with equal properties, implementation of ModifiableRootModel.removeOrderEntry in the old model removes the specified instance only, where as implementation in the new model removed all of them. So MavenRootModelAdapterLegacyImpl::initOrderEntries removed all copies after calling removeOrderEntry for the first instance, and failed with "does not belong to modifiableRootModel" error when trying to remove remaining copies. Also in UI it looks strange if removing one item in the dependencies list removes other copies automatically. So now ModifiableRootModel.removeOrderEntry removes the specified entry only in the new implementation as well.

GitOrigin-RevId: 54cdce78b27134988e9b68c393f4985d977dfa1d