[groovy] add `equals`/`hashCode` to `GrLightMethodBuilder`; this fixed generics inference in light methods
What was happening:
Several threads produced different `MethodResolveResult`-s for the same reference, one of results was cached.
Highlighting uses resolve result computed in this thread; this resolve result might be not cached.
When time comes, the current result infers its substitutor, and during inference another result is returned from
`GroovyReference#resolve()`; the inference fails because the results are not equal, but the highlighting is still
applied as if the current result was cached.

This fixes `org.jetbrains.plugins.groovy.builder.JsonBuilderTest#test highlighting` and possibly some idempotence
checker errors.

GitOrigin-RevId: c608f42b36aaea58b0a12f70d7499b6bef3abb2d