when asked for an injected caret, do not return host caret
because it would lead to a great confusion otherwise, if one editor is embedded (swing-speaking)
into another, like GridCellEditorComponent into host EditorImpl.
In this case DataManager used to start looking up caret in the bottom editor GridCellEditorComponent,
failed to find injected caret there,
went to parent editor,
asked for injected caret but got the host one instead,
and every request to the resulting caret was garbled after that.

GitOrigin-RevId: acd4ea34a23d895863e85cf979c18ab3f3127c96