[workspace model] IDEA-268589 Fix update project dependencies for external system projects
The root cause of the problem was with the project library which was added e.g at previous project import
and during the next import it was removed but the modifiable for it still exists at
`IdeModifiableModelsProvider.myModifiableLibraryModels` thus such modifiable model should be detected
and disposed. Otherwise any operations with such modifiable model will lead to the exception because
library was already removed from the store.

GitOrigin-RevId: c816a4f6866d0d3379dea6ac8f541078bf3345b1