[workspace model] IDEA-268254 Cleanup entities references if during rBs we get the same but with different entity source
For some reasons it's possible that at replace by source we can get the same entity but with different entity source
and for such cases the store should stay at the consistent state after rBs in terms of references to this such entity.
The main scenario which was at this ticket:
1) Store contains `LibraryEntity` with Persistent Id = "XStream" and entity source FileInDirectory and reference to
`LibraryPropertiesEntity` which doesn't have Persistent Id.
2) After some modification **(this should be further investigated)** at `JpsProjectSerializersImpl.reloadFromChangedFiles`
we got a buildr with the same entities for "XStream" lib but with different entity source
3) After rBs we got store with one `LibraryEntity` with Persistent Id = "XStream" (because it has persistent id and
we know what to replace) and two `LibraryPropertiesEntity` for the same lib with the relationship one-to-one
between this entities (one of them was from the old store because it has entity source which doesn't match the rBs).
At such scenario rBs should remove all references from the store for the replaced entity to keep the store at the
consistent state

GitOrigin-RevId: 090e66592919589adc09ff7104bb90c8e7a5ce4e