Extract database.connectivity from database.impl (LAB-77)
database.connectivity is a module that allows to connect to a database. The module is independent from the Datagrip UI and contains only the core database functionality.
The tests are located in the corresponding module database.connectivity-tests. The xml configuration for the UI independent test suit is located in codeServer.db.connectivity module.

This is the initial commit.

GitOrigin-RevId: d91fbd1049a0a68ac1d2e633d9994c452ad07b84