[workspace model] IDEA-CR-71444 Introduce method for making store at the consistent state where possible for one to one refs
How this method works: we try to find all 1-1 references where current entity is parent or child. We check that
such reference also exist at the replacing store. And if such reference exist at the second store we can safely
delete related by the reference entity because we sure that the reference will be restored and entity from the
second store will be added to the main.
One important notes of the entity remove: we make it only if the second store contains entity which matched by
source filter, we don't remove entity which already was replaced and we don't remove the the object itself.

GitOrigin-RevId: ec8418269bacc8e229cd70ed48a6dd0ce817321f