[build scripts] include sources of libraries into product sources zip archive (IDEA-168113)
If we pack sources of all libraries used from the modules included into the product, it'll greatly increase size of the archive (from current 65Mb to 177Mb). So here we pack source of libraries which are included into compilation classpath for platform API modules, this way we'll get sources of all libraries useful for plugin developers and size of the archive will be reasonable (85Mb).

GitOrigin-RevId: 55eea9d4acdecce3f0ccab6025d058fff0e8b783