Share LaFManagerImpl settings between different OSs: set the roaming type to DEFAULT
Currently most of themes (including Darcula and IntelliJ Light which cover most of all theme usages) are os-independent, so it is expected to synchronize them between different computers, even if there are different OSs on them.

There are two exceptions: Windows 10 Light and macOS Light themes, which are OS-dependent. However, both fallback to IntelliJ Light if they are not available in the current IDEA, and the fallback doesn't modify

GitOrigin-RevId: c56133cc97945f71a3e562d6873f49f96cb6b9ab