Fallback to IntelliJ Light if native light themes are loaded from laf.xml but don't exist on this computer
This may happen e.g. if a user uses IDE Settings Sync to synchronize between Mac and Windows. It was decided that synchronizing of LaFs should be os-independent since there are only two os-specific themes (and both are actually obsolete).

Previously it would fallback to Darcula which is not convenient, since both native themes are light.

GitOrigin-RevId: 3a43cf27e2f078e891a1548645b9d8cfe1c757b8