[git-index] cancel refresh tasks on project closing
Because "invokeAndWait" method is called with the default modality, "Executor#awaitTermination" method may not finish waiting as "invokeAndWait" won't be able to run during modal progress. Currently it does not seem to be a big problem, since it is important only to wait for git tasks to finish (or cancel them). However it makes sense to leave waiting duration short.

GitOrigin-RevId: 7cc100c6f6d913c5fc017cdf377c5a85d570f0b5