[freeze][dataflow from here] IDEA-268322 Freeze when Analyze Data Flow from Here
This patch restores the blocking read computation in JavaOverridingMethodsSearcher because the "nonBlocking()" call changes semantics of JavaOverridingMethodSearcher. Nonblocking made it be interrupted by write actions and everything became invalid in the middle of a computation, which might have affected the clients.

The analyze from/to here panel should be rewritten to the async model with the help from AsyncTreeModel.

GitOrigin-RevId: ed6864ba098724487f9836ecf880bbed60da0545