[workspace model] fix removal of manually added facet in module imported from external system (IDEA-269842)
When a FacetEntity which ModuleEntity is stored in external storage is removed, the corresponding FileInDirectory became obsolete, but its file wasn't removed because its JpsImportedEntitySource wasn't present in affected sources. Now we will add such JpsImportedEntitySource into entitiesToSave map. This way we'll save entities to external storage if entities saved in the corresponding internal source are changed, but performance effect of this should be negligible.

Also module in regularFacetInImportedModule test data was renamed to match other test data, this allows us to reuse it in the test.

GitOrigin-RevId: 289b35e99df4e43c4c74036920d18b6ec51f15a2