show embedded progress icon in gutter component
Two implementations:
1. sync case for popup menus (e.g. "Run ..." gutter icon)
2. async case for `NavigationGutterIconRenderer`

Note that sync implementations can also be extended to support async case - a "progress remover" can be passed to async code via any static variable. Both approaches for async case require client code modifications, so I decided to use "renderer icon" approach to demonstrate that it is also possible and because it is a bit smaller.

GitOrigin-RevId: 809ca4da56b93e5d985fdb46c51b8216a825f028