IJP-630 (wip) 'Actions on Save' settings page: sync state of the 'action on save enabled' check boxes on 2 pages.
Some 'actions on save' can be configured in 2 places: on the 'Actions on Save' page and on some other technology-specific page in
Settings (Preferences). For example, 'Run eslint --fis on save': there's a check box on the Actions on Save page and on the ESLint
page in Settings. The state of these two check boxes should be the same.

GitOrigin-RevId: aae5e6807b1eb4dc9f5be114965480764f3334ce