IJP-630 (wip) 'Actions on Save' settings page: ensure Settings dialog knows that SomeOtherConfigurable becomes modified
when user changes something on the 'Actions on Save' page. Some 'actions on save' can be configured in 2 places: on the 'Actions on Save' page and on
some other technology-specific page in Settings (Preferences). For example, 'Run eslint --fis on save': there's a check box on the Actions on Save
page and on the ESLint page in Settings. When user toggles 'Run eslint --fis on save' check box on the 'Actions on Save' page, 'ESLint' age should
update its 'modified' state.

GitOrigin-RevId: 341357b9f72b3bd5839109defe2320cd899b3148