[ui] remove useless components from AbstractPopup
"myBottomComponent" was not used at all, "myBottomPanel" was added to "myContent", replacing "myAdComponent". This prevented displaying ad text when it was passed to the popup in the AbstractPopup#init method (this is done by ComponentPopupBuilderImpl, for example).

follow-up: c0efb4a95443bedd13e6d52cbf2d6ad02ea3fb17

IDEA-267018

GitOrigin-RevId: 3e6cb3a8f23e0e84fc8d3b26c45b06e3ba00efdc