[properties] IDEA-266551 Properties files opening with UTF-8 instead of default encoding since 2021.1
The default encoding for properties files is ISO-8859-1 but IDEA used to take the default system local if the encoding for properties files wasn't set explicitly, which might be either UTF-8 or windows-1252 or something else. This patch defaults the encoding for properties files to ISO-8859-1 and fixes messages in the File Encodings settings panel.

Signed-off-by: Nikita Eshkeev <nikita.eshkeev@jetbrains.com>

GitOrigin-RevId: 26e9ef5cc03b43ecd2e47354f626cb4f4a77d474