Don't share paths defined in PathMacrosImpl
Paths are workstation-specific and shouldn't be shared. It is quite probable that a user sets up a similar environment on both computers, so sharing would be useful for them. But at the same time if the users doesn't have a common environment, then unexpected synchronization would be very annoying (even if it can be turned off).

In addition to the general product decision of not syncing paths, the PathMacrosImpl is a special component (used for initialization of other components). Its path is hardcoded and referenced in many other places, so fixing it after changing the layout of per-os settings would require another round of error-prone hardcoding.

IDEA-152940

GitOrigin-RevId: af33d0e915362514d47f56055f9d04c24d86b404