Don't share paths defined in application libraries
Paths are workstation-specific and shouldn't be shared. It is quite probable that a user sets up a similar environment on both computers, so sharing would be useful for them. But at the same time if the users doesn't have a common environment, then unexpected synchronization would be very annoying (even if it can be turned off).

IDEA-152940

GitOrigin-RevId: ccc6d511dae74c684355bb295df0bd76b4ecc0df