[vcs][properties] IDEA-270602 Optimize precommit code analysis for property files
Put a closure that extracts psiElements that are being committed in the userData of a file here before the commit code smell analyzer starts and use the closure to analyze only the extracted properties in the UnusedPropertyInspection inspection. This behaviour is controlled by the "vcs.code.analysis.before.checkin.check.unused.only.changed.properties" registry key as it may be unexpected that not all properties are analyzed.

GitOrigin-RevId: f30dc85847fc83205428c1a026f9cdd4a3920872