Update shortcuts for conflicting actions in Diff.
After moving diffs to the editor, shortcuts in Diff began to conflict with shortcuts used to manage tabs. (Next/previos tab and Switcher)
Changed the less used shortcuts.

Found new shortcuts for Accept Left/Right.
Set Shift+Ctrl+Tab to Switch focus between Diff sides. This shortcut is also used by switcher for reverse navigation. However, it only collides when the focus is in diff, but works fine when the Switcher is opened, which seems acceptable, as Ctrl+Tab is handier to open the Switcher.

Fixes IDEA-239744 and IDEA-237547

Review IJ-MR-9945

GitOrigin-RevId: 4c7f640e34e950283a11d5eaa9f034eac80cb4ea