KTIJ-5114: UselessCall inspection for Sequence
Updated UselessCallOnCollectionInspection to extend map/filter call
reduction to Sequence<>.

There is a limitation: kotlin.io library contains the FileTreeWalk
class defined as a subclass of Sequence<File>, but the inspection currently
skips method calls for non-generic classes. As a result, no fix is proposed
for FileTreeWalk.mapNotNull etc. I am planning to investigate this case
separately from this commit.

^KTIJ-5114 Fixed

GitOrigin-RevId: ccaa5e647e7e7d96ed588e5716342851f79b7376