[git-index] avoid acquiring read lock under potemkin progress
GitIndexFileSystemRefresher writes GitIndexVirtualFile content by running git in edt under write lock. In order to avoid freezes, PotemkinProgress is used for this task. However, for providing environment variables to start git command PyVirtualEnvVcsCustomizer needs to use read lock to retrieve python sdk instance. This results in a deadlock. In order to temporarily fix the problem, this commit adds a hack: a check that skips calling environment variables customizers altogether under PotemkinProgress, thus avoiding the read lock. This will result in running git commands for writing to the index without environment variables to activate virtualenv or conda. This hack is to be removed later when a more permanent solution is provided.

IDEA-266329

GitOrigin-RevId: 2e1821656fc3ea613581545de601f8d0f15c384b