[test framework] fix registration and disposal of SDK for light tests (IDEA-CR-71564)
JDKs need to be registered in ProjectJdkTable before we can use them in modules, so now we register SDK before the project is initialized and use a separate disposable for it to unregister it after disposing project-related things to avoid assertions. Also there is no need to unregister all JDKs after LightPlatformTestCase is finished (SdkLeakTracker does this anyway).

GitOrigin-RevId: 0afca3954cd1897ea452584a00d87aa9d21f7447