[workspace model] properly update *.iml and modules.xml when module configuration is moved to external storage
It may happen that an entity source for a ModuleEntity was changed to external storage, but there are other entities (facets) which are still stored in *.iml file, so it won't become obsolete. In such cases we need to remove obsolete configuration tags from *.iml file and remove reference to it from .idea/modules.xml to avoid having duplicating data which may lead to problems (IDEA-271172).

GitOrigin-RevId: 3c504e2f471392a4d3b120daaa2ac09f621a58a7