Injections: making `EditorWindowTrackerImpl.disposeEditorFor` dispose all matching editors
instead of only first one, which in some cases even doesn't contain the given `DocumentWindowImpl` when the required editor (with given DocumentWindowImpl) could appear later in the list.

There is no guarantee that only one editor will be created (because multiple DocumentWindowImpl could be created by parallel threads), so we are disposing them all.

GitOrigin-RevId: 77f50044afeeef6e2c9c7ddb0623fa475bb60d85