[vcs-log] allow to call AbstractDataGetter#getCommitData from background thread
Normally AbstractDataGetter should be used from the event dispatch thread, however after actions update was moved to background thread getCommitData started to get called from background which lead to missing cached commits (EA-256553). The cache itself is thread-safe, but the code that handles loading is not. This commit adds a hack that skips loading commits when getCommitData is called from background: only cached details are returned or a placeholder is created when commit is not in cache.

EA-256553, EA-286638, EA-264456, EA-258237

GitOrigin-RevId: 02735add379ca8c4fc42a9eb782a71e6770381b3