settings-repository: don't resolve deprecated per-os storage to file in repository: IDEA-272508
Per-os component has two storages, one is actual e.g. "mac/keymap.xml", another one is deprecated e.g. "keymap.xml". The actual storage is used to read settings from it and write settings to it, so it should be resolved to settings-repository's real file located at "_mac/keymap.xml". The deprecated storage is not used for read/write, but is deleted when the component is saved, and this deletion request is passed to the Settings Repository's StreamProvider.

Previously the IcsStreamProvider would resolve "keymap.xml" to "_mac/keymap.xml" as well, so deletion of the deprecated "keymap.xml" storage resulted in deletion of the real actual "_mac/keymap.xml". The fix is not to resolve per-os storages without os-prefix.

(cherry picked from commit cba40583a86ce5b9ebb22ff280269eacc940ef3c)

GitOrigin-RevId: 1b786f4bede62fc601d1c6dcbefe926f7ddc8cd1